Discovered problem in Titanium wasn't fn.apply
After further investigation the issue was that the argument being passed
into handler.proxy wasn't an array when passed by Titanium's
'fireEvent'. A simple check to convert it to an array if needed was all
that was necessary to restore use of 'apply' and remove the splat
argument to 'call'.