Fixed bug when using 'class' attributes in TI
Because 'class' is a reserved word in Javascript, Coffeescript helpfully
converts hash keys named 'class' from hash.class syntax to hash['class']
syntax. While this works in most Javascript environments, in Titanium an
error is thrown when trying to access an array-like ([]) accessor on a
Titanium view.

This patch converts 'class' attributes to '_class' when set via .attr.

* Note: One possibility to work around this was to use `className`
instead of `class`. The problem with this is that Titanium treats
className differently on TableViewRows and makes assumptions about the
similarity of rows with matching classNames that may not hold true.