Updated traversal tests
The previous test was using a Tab as an arbitrary view name in a find()
test. This will no longer be valid after we add error checking to the
append method to only allow tabs to be added to tab groups. A ScrollView
serves the purpose well because it can be added to a parent View and can
contain the usual types of children.