Removed backbone-extensions for API consistency
This removes the tagName -> viewName alias in favor of just using
tagName (like Backbone on the web does). With the upcoming XML-based
element creation syntax, tagName makes more conceptual sense. We also
removed the override of the default tagName to 'View'. Now, if a view
has a tagName of 'div' then it is renamed to 'View' before creation.
This is surprisingly helpful when using other libraries, especially Jade
in combination with XML-based element creation (you'll see!).

It also removes the override of _configure that merged attributes passed
in the constuctor. As we move towards a more CSS-like approach for
styling elements this should provide less of a benefit, so I'm removing
it in favor of maintaining consistency with Backbone.

Now the only thing needed to mix titanium-backbone in to Backbone is:

Backbone.$ = require('tb').$