Updated mock TableView to more closely follow Ti
The biggest change is that Titanium does not allow direct access to
TableView.rows. The rows are stored under TableView.data but if a row is
added via appendRow then a default section is created and the row is
added under that section. In that case the table rows are available
under table.data[0].rows.